+++ CHANGES.md
@@ -5,6 +5,7 @@ Changes in 10.4
 ---------------
 - Modcat is addon module, set custom module icon in Profiles.php & Exceptions.php
 - SQL performance: use NOT EXISTS instead of NOT IN + LIMIT 1000 in Portal.php
+- Add student name to Student's Absences and Grades delete prompt in MassDrops.php
 
 Changes in 10.3.2
 -----------------